Rename Parser/Parser -> Parser/OutputParser
